`no-noise-split-diff-view` --> `no-unnecessary-split-diff-view`
